Refactor AuthorizeStakingApps component
